Add rake task that can be run as a cron job to periodically index content from aspace
