+++ app/controllers/pipette/process_finding_aids_controller.rb
@@ -2,6 +2,8 @@
 
 module Pipette
   class ProcessFindingAidsController < ApplicationController
+    include Pipette::AspaceHelper
+
     def process_all_ead
       ead_ids = get_aspace_ids(params[:index_since])
       if ead_ids.empty?
@@ -45,39 +47,6 @@ module Pipette
 
       flash[:notice] = "#{aspace_id} deleted from ArCHy"
       redirect_to job_status_index_path
-    end
-
-    private
-
-    def get_aspace_ids(index_since)
-      query_vars = {
-        all_ids: true
-      }
-      query_vars[:modified_since] = time_offset unless index_since.nil?
-
-      Pipette::AspaceClient.client.get('resources', { query: query_vars }).parsed
-    end
-
-    # Time periods are in seconds
-    def time_offset
-      offset = params[:index_since].downcase
-      return nil if offset == 'all'
-
-      case offset
-      when 'hour'
-        updated_since(3600)
-      when 'half-day'
-        updated_since(43_200)
-      when 'full-day'
-        updated_since(86_400)
-      else # one week
-        updated_since(604_800)
-      end
-    end
-
-    # Get the number of seconds since a time offset
-    def updated_since(period)
-      (Time.parse(Time.now.to_s).to_i - period).to_s
     end
   end
 end
