--- lib/tasks/pipette_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :pipette do
-#   # Task goes here
-# end
